requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
python-dotenv